+++ handler.js
@@ -1,14 +1,9 @@
 'use strict';
 
 const queryString = require('query-string');
-const _ = require('lodash');
 
-const {
-  saveTask,
-  findTaskById,
-  getExpiringTasksIn24Hrs,
-} = require('./taskManager');
-const { slackPublish } = require('./slackPublisher');
+const { slackPublish } = require('./services/slackPublisher');
+const { saveATask, getATask, getAllPendingTasks } = require('./services/tasks');
 
 const createResponse = (statusCode, message) => ({
   statusCode,
@@ -29,9 +24,6 @@ module.exports.hello = (event, context, callback) => {
       input: event,
     }),
   );
-
-  // Use this code if you don't use the http event with the LAMBDA-PROXY integration
-  // return { message: 'Go Serverless v1.0! Your function executed successfully!', event };
 };
 
 module.exports.post = (event, context, callback) => {
@@ -59,54 +51,6 @@ module.exports.createTask = (event, context, callback) => {
   callback(null, response);
 };
 
-const saveATask = (text, userId, userName, callback) => {
-  const textArray = text.split(',');
-  const taskTitle = textArray[0];
-  const dueDate = textArray[1];
-
-  const taskDescription = _.tail(_.tail(textArray)).join();
-
-  console.log('taskTitle: ', taskTitle);
-  console.log('dueDate: ', dueDate);
-  console.log('taskDescription: ', taskDescription);
-  console.log(`userId: ${userId} userName: ${userName}`);
-
-  saveTask(taskTitle, taskDescription, dueDate, userId, userName)
-    .then(res => {
-      console.log('response: ', res);
-      callback(null, createResponse(200, res));
-    })
-    .catch(err => {
-      console.log(err);
-      callback(null, createResponse(500, 'Error on saving task'));
-    });
-};
-
-const getATask = (text, callback) => {
-  console.log('text in request body:', text);
-
-  findTaskById(text)
-    .then(result => {
-      console.log('result: ', result);
-      callback(null, createResponse(200, result));
-    })
-    .catch(error => {
-      console.log(error);
-      callback(null, createResponse(500, 'Error getting a task'));
-    });
-};
-
-const getAllPendingTasks = callback => {
-  getExpiringTasksIn24Hrs()
-    .then(result => {
-      console.log(result);
-      callback(null, createResponse(200, { attachments: result }));
-    })
-    .catch(error => {
-      console.log(error);
-      callback(null, createResponse(500, 'Error getting all pending tasks'));
-    });
-};
 
 module.exports.tasks = (event, context, callback) => {
   const {
@@ -117,12 +61,16 @@ module.exports.tasks = (event, context, callback) => {
   } = queryString.parse(decodeURIComponent(event.body));
 
   console.log('command: ', command);
+
   if (command === '/new-task') {
     saveATask(text, userId, userName, callback);
+
   } else if (command === '/find-task') {
     getATask(text, callback);
+
   } else if (command === '/pending-tasks') {
     getAllPendingTasks(callback);
+
   } else {
     callback(
       null,
