+++ services/tasksFormatter.js
@@ -59,28 +59,6 @@ module.exports.formatAttachment = tasksList => {
   console.log('task List inside formatAttachment: ', tasksList);
   const attachments = tasksList.map((task, i) => {
     return formatter(task, i);
-    // return {
-    //   fallback: task?.taskDescription || '-',
-    //   color: '#333333',
-    //   title: task?.taskTitle || 'no title',
-    //   text: task?.taskDescription || 'no text',
-    //   fields: [
-    //     {
-    //       title: 'Task Id',
-    //       value: task?.taskId || 'wrong id, please message to support',
-    //     },
-    //     {
-    //       title: 'Due date',
-    //       value:
-    //         task?.dueDate &&
-    //         format(new Date(task?.dueDate), 'yyyy-MM-dd HH:mm'),
-    //     },
-    //     {
-    //       title: 'Author',
-    //       value: task?.username && '-',
-    //     },
-    //   ],
-    // };
   });
 
   console.log('attachments: ', attachments);
@@ -94,28 +72,3 @@ module.exports.formatAttachmentSingleTask = task => {
   console.log('attachments: ', attachments);
   return attachments;
 };
-
-/*
-{
-    "attachments": [
-        {
-            "color": "#333333",
-            "author_name": "Username",
-            "title": "Task title",
-            "text": "Task description",
-            "fields": [
-                {
-                    "title": "Task id",
-                    "value": "TaskId"
-                },
-				{
-                    "title": "Due date",
-                    "value": "Duedate"
-                }
-            ],
-            "footer": "Learning serverless tasks manager",
-            "ts": 123456789
-        }
-    ]
-}
-*/
